requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.10
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.14.0
duplicity==0.8.21
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
exceptiongroup==1.2.0
fastapi==0.110.1
fasteners==0.14.1
filelock==3.6.0
fonttools==4.49.0
future==0.18.2
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Keras-Applications==1.0.8
keras-ocr==0.9.3
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.8.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
packaging==21.3
pandas==2.2.1
paramiko==2.9.3
pastel==0.2.1
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.4.0
poetry==1.1.12
poetry-core==1.0.7
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pyclipper==1.3.0.post5
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.10
pytest==7.4.4
pytest-django==4.8.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
scikit-image==0.22.0
scipy==1.12.0
SecretStorage==3.3.1
shapely==2.0.3
shellingham==1.4.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
starlette==0.37.2
systemd-python==234
tifffile==2024.2.12
tomli==2.0.1
tomlkit==0.9.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.29.0
validators==0.22.0
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
